Restructure files, remove status-bar related stuff
