+++ Sources/Simctl/HttpMethod.swift
@@ -1,0 +1,9 @@
+
+
+@usableFromInline
+internal enum HttpMethod: String {
+    case get = "GET"
+    case post = "POST"
+    case put = "PUT"
+}
+extension HttpMethod: Equatable { }
