+++ Sources/Simctl/ProcessEnvironmentKey.swift
@@ -1,0 +1,27 @@
+
+internal enum ProcessEnvironmentKey: String {
+    case simulatorAudioDevicesPlistPath = "SIMULATOR_AUDIO_DEVICES_PLIST_PATH"
+    case simulatorAudioSettingsPath = "SIMULATOR_AUDIO_SETTINGS_PATH"
+    case simulatorBootTime = "SIMULATOR_BOOT_TIME"
+    case simulatorCapabilities = "SIMULATOR_CAPABILITIES"
+    case simulatorDeviceName = "SIMULATOR_DEVICE_NAME"
+    case simulatorExtendedDisplayProperties = "SIMULATOR_EXTENDED_DISPLAY_PROPERTIES"
+    case simulatorFramebufferFramework = "SIMULATOR_FRAMEBUFFER_FRAMEWORK"
+    case simulatorHIDSystemManager = "SIMULATOR_HID_SYSTEM_MANAGER"
+    case simulatorHostHome = "SIMULATOR_HOST_HOME"
+    case simulatorLegacyAssetSuffic = "SIMULATOR_LEGACY_ASSET_SUFFIX"
+    case simulatorLogRoot = "SIMULATOR_LOG_ROOT"
+    case simulatorMainScreenHeight = "SIMULATOR_MAINSCREEN_HEIGHT"
+    case simulatorMainScreenPitch = "SIMULATOR_MAINSCREEN_PITCH"
+    case simulatorMainScreenScale = "SIMULATOR_MAINSCREEN_SCALE"
+    case simulatorMainScreenWidth = "SIMULATOR_MAINSCREEN_WIDTH"
+    case simulatorMemoryWarnings = "SIMULATOR_MEMORY_WARNINGS"
+    case simulatorModelIdentifier = "SIMULATOR_MODEL_IDENTIFIER"
+    case simulatorProductClass = "SIMULATOR_PRODUCT_CLASS"
+    case simulatorRoot = "SIMULATOR_ROOT"
+    case simulatorRuntimeBuildVersion = "SIMULATOR_RUNTIME_BUILD_VERSION"
+    case simulatorRuntimeVersion = "SIMULATOR_RUNTIME_VERSION"
+    case simulatorSharedResourcesDirectory = "SIMULATOR_SHARED_RESOURCES_DIRECTORY"
+    case simulatorUdid = "SIMULATOR_UDID"
+    case simulatorVersionInfo = "SIMULATOR_VERSION_INFO"
+}
