+++ Sources/Simctl/SimctlClientEnvironment.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public protocol SimctlClientEnvironment {
+  var host: SimctlClient.Host { get }
+  var bundleIdentifier: String? { get }
+  var deviceUdid: UUID { get }
+}
