+++ Sources/SimctlCLI/ListDevices.swift
@@ -12,27 +12,27 @@ import SimctlShared
 import Swifter
 
 public enum ListDevicesError: Swift.Error {
-    case dataConversionFailed
+  case dataConversionFailed
 }
 
 struct ListDevices: ParsableCommand {
-    static var configuration = CommandConfiguration(abstract: "List the simulator devices")
-
-    mutating func run() throws {
-        print("\(listDevices().map { $0.description }.sorted().joined(separator: "\n"))")
-    }
+  static var configuration = CommandConfiguration(abstract: "List the simulator devices")
+  
+  mutating func run() throws {
+    print("\(listDevices().map { $0.description }.sorted().joined(separator: "\n"))")
+  }
 }
 
 private func listDevices() -> [SimulatorDevice] {
-    do {
-        let devicesJSONString = try shellOut(to: .simctlList(.devices, true))
-        guard let devicesData: Data = devicesJSONString.data(using: .utf8) else {
-            throw ListDevicesError.dataConversionFailed
-        }
-        let decoder = JSONDecoder()
-        let listing = try decoder.decode(SimulatorDeviceListing.self, from: devicesData)
-        return listing.devices
-    } catch {
-        return []
+  do {
+    let devicesJSONString = try shellOut(to: .simctlList(.devices, true))
+    guard let devicesData: Data = devicesJSONString.data(using: .utf8) else {
+      throw ListDevicesError.dataConversionFailed
     }
+    let decoder = JSONDecoder()
+    let listing = try decoder.decode(SimulatorDeviceListing.self, from: devicesData)
+    return listing.devices
+  } catch {
+    return []
+  }
 }
