+++ Sources/SimctlCLI/SimctlServer.swift
@@ -12,360 +12,304 @@ import Swifter
 /// The server used to receive commands from an app and
 /// translate them into commands on the local machine.
 internal final class SimctlServer {
-    let server: HttpServer
+  let server: HttpServer
 
-    init() {
-        server = HttpServer()
+  init() {
+    server = HttpServer()
+  }
+
+  /// Start a server that listens to library requests from your app and executes simctl commands on your machine.
+  /// - Parameter port: the port on which to listen.
+  func startServer(on port: SimctlShared.Port) {
+    do {
+      try server.start(port)
+      print("Server listening on port \(port)...")
+      RunLoop.main.run()
+    } catch {
+      fatalError("Unable to start server on port \(port)")
     }
+  }
 
-    /// Start a server that listens to library requests from your app and executes simctl commands on your machine.
-    /// - Parameter port: the port on which to listen.
-    func startServer(on port: SimctlShared.Port) {
-        do {
-            try server.start(port)
-            print("Server listening on port \(port)...")
-            RunLoop.main.run()
-        } catch {
-            fatalError("Unable to start server on port \(port)")
+  /// Stop the server.
+  func stop() {
+    server.stop()
+  }
+
+  /// Callback to be executed on push notifcation send request.
+  /// - Parameter closure: The closure to be executed.
+  func onPushNotification(_ closure: @escaping (UUID, String?, PushNotificationContent) -> Result<String, Swift.Error>) {
+    server.POST[ServerPath.pushNotification.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
+
+      guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
+        return .badRequest(.text("Bundle Id missing or corrupt."))
+      }
+
+      let bodyData = Data(request.body)
+      let decoder = JSONDecoder()
+      do {
+        let pushContent: PushNotificationContent = try decoder.decode(PushNotificationContent.self, from: bodyData)
+
+        let result = closure(deviceId, bundleId, pushContent)
+
+        switch result {
+        case let .success(output):
+          return .ok(.text(output))
+
+        case let .failure(error):
+          return .badRequest(.text(error.localizedDescription))
         }
+      } catch {
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
+  }
 
-    /// Stop the server.
-    func stop() {
-        server.stop()
+  /// Callback to be executed on privacy change request.
+  /// - Parameter closure: The closure to be executed.
+  func onPrivacy(_ closure: @escaping (UUID, String?, PrivacyAction, PrivacyService) -> Result<String, Swift.Error>) {
+    server.GET[ServerPath.privacy.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
+
+      guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
+        return .badRequest(.text("Bundle Id missing or corrupt."))
+      }
+
+      guard let action: PrivacyAction = request.headerValue(for: .privacyAction) else {
+        return .badRequest(.text("Privacy action missing or corrupt."))
+      }
+
+      guard let service: PrivacyService = request.headerValue(for: .privacyService) else {
+        return .badRequest(.text("Privacy service missing or corrupt."))
+      }
+
+      let result = closure(deviceId, bundleId, action, service)
+
+      switch result {
+      case let .success(output):
+        return .ok(.text(output))
+
+      case let .failure(error):
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
+  }
 
-    /// Callback to be executed on push notifcation send request.
-    /// - Parameter closure: The closure to be executed.
-    func onPushNotification(_ closure: @escaping (UUID, String?, PushNotificationContent) -> Result<String, Swift.Error>) {
-        server.POST[ServerPath.pushNotification.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
+  /// Callback to be executed on rename device request.
+  /// - Parameter closure: The closure to be executed.
+  func onRename(_ closure: @escaping (UUID, String?, String) -> Result<String, Swift.Error>) {
+    server.GET[ServerPath.renameDevice.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
 
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
+      guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
+        return .badRequest(.text("Bundle Id missing or corrupt."))
+      }
 
-            let bodyData = Data(request.body)
-            let decoder = JSONDecoder()
-            do {
-                let pushContent: PushNotificationContent = try decoder.decode(PushNotificationContent.self, from: bodyData)
+      guard let deviceName: String = request.headerValue(for: .deviceName) else {
+        return .badRequest(.text("No device name parameter provided."))
+      }
 
-                let result = closure(deviceId, bundleId, pushContent)
+      let result = closure(deviceId, bundleId, deviceName)
 
-                switch result {
-                case let .success(output):
-                    return .ok(.text(output))
+      switch result {
+      case let .success(output):
+        return .ok(.text(output))
 
-                case let .failure(error):
-                    return .badRequest(.text(error.localizedDescription))
-                }
-            } catch {
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
+      case let .failure(error):
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
+  }
 
-    /// Callback to be executed on privacy change request.
-    /// - Parameter closure: The closure to be executed.
-    func onPrivacy(_ closure: @escaping (UUID, String?, PrivacyAction, PrivacyService) -> Result<String, Swift.Error>) {
-        server.GET[ServerPath.privacy.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
+  /// Callback to be executed on terminate app device request.
+  /// - Parameter closure: The closure to be executed.
+  func onTerminateApp(_ closure: @escaping (UUID, String?, String) -> Result<String, Swift.Error>) {
+    server.GET[ServerPath.terminateApp.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
 
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
+      guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
+        return .badRequest(.text("Bundle Id missing or corrupt."))
+      }
 
-            guard let action: PrivacyAction = request.headerValue(for: .privacyAction) else {
-                return .badRequest(.text("Privacy action missing or corrupt."))
-            }
+      guard let targetAppBundleId: String = request.headerValue(for: .targetBundleIdentifier) else {
+        return .badRequest(.text("No target app bundle id parameter provided."))
+      }
 
-            guard let service: PrivacyService = request.headerValue(for: .privacyService) else {
-                return .badRequest(.text("Privacy service missing or corrupt."))
-            }
+      let result = closure(deviceId, bundleId, targetAppBundleId)
 
-            let result = closure(deviceId, bundleId, action, service)
+      switch result {
+      case let .success(output):
+        return .ok(.text(output))
 
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
-
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
+      case let .failure(error):
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
+  }
 
-    /// Callback to be executed on rename device request.
-    /// - Parameter closure: The closure to be executed.
-    func onRename(_ closure: @escaping (UUID, String?, String) -> Result<String, Swift.Error>) {
-        server.GET[ServerPath.renameDevice.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
+  func onErase(_ closure: @escaping (UUID) -> Result<String, Swift.Error>) {
+    server.GET[ServerPath.erase.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
 
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
+      let result = closure(deviceId)
 
-            guard let deviceName: String = request.headerValue(for: .deviceName) else {
-                return .badRequest(.text("No device name parameter provided."))
-            }
+      switch result {
+      case let .success(output):
+        return .ok(.text(output))
 
-            let result = closure(deviceId, bundleId, deviceName)
-
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
-
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
+      case let .failure(error):
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
+  }
 
-    /// Callback to be executed on terminate app device request.
-    /// - Parameter closure: The closure to be executed.
-    func onTerminateApp(_ closure: @escaping (UUID, String?, String) -> Result<String, Swift.Error>) {
-        server.GET[ServerPath.terminateApp.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
+  func onSetDeviceAppearance(_ closure: @escaping (UUID, String?, DeviceAppearance) -> Result<String, Swift.Error>) {
+    server.GET[ServerPath.deviceAppearance.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
 
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
+      guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
+        return .badRequest(.text("Bundle Id missing or corrupt."))
+      }
 
-            guard let targetAppBundleId: String = request.headerValue(for: .targetBundleIdentifier) else {
-                return .badRequest(.text("No target app bundle id parameter provided."))
-            }
+      guard let appearance: DeviceAppearance = request.headerValue(for: .deviceAppearance) else {
+        return .badRequest(.text("Device appearance missing or corrupt."))
+      }
 
-            let result = closure(deviceId, bundleId, targetAppBundleId)
+      let result = closure(deviceId, bundleId, appearance)
 
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
+      switch result {
+      case let .success(output):
+        return .ok(.text(output))
 
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
+      case let .failure(error):
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
+  }
 
-    func onErase(_ closure: @escaping (UUID) -> Result<String, Swift.Error>) {
-        server.GET[ServerPath.erase.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
+  func onTriggerICloudSync(_ closure: @escaping (UUID, String?) -> Result<String, Swift.Error>) {
+    server.GET[ServerPath.iCloudSync.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
 
-            let result = closure(deviceId)
+      guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
+        return .badRequest(.text("Bundle Id missing or corrupt."))
+      }
 
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
+      let result = closure(deviceId, bundleId)
 
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
+      switch result {
+      case let .success(output):
+        return .ok(.text(output))
+
+      case let .failure(error):
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
+  }
 
-    func onSetDeviceAppearance(_ closure: @escaping (UUID, String?, DeviceAppearance) -> Result<String, Swift.Error>) {
-        server.GET[ServerPath.deviceAppearance.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
+  /// Callback to be executed on uninstall app device request.
+  /// - Parameter closure: The closure to be executed.
+  func onUninstallApp(_ closure: @escaping (UUID, String?, String) -> Result<String, Swift.Error>) {
+    server.GET[ServerPath.uninstallApp.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
 
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
+      guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
+        return .badRequest(.text("Bundle Id missing or corrupt."))
+      }
 
-            guard let appearance: DeviceAppearance = request.headerValue(for: .deviceAppearance) else {
-                return .badRequest(.text("Device appearance missing or corrupt."))
-            }
+      guard let targetAppBundleId: String = request.headerValue(for: .targetBundleIdentifier) else {
+        return .badRequest(.text("No target app bundle id parameter provided."))
+      }
 
-            let result = closure(deviceId, bundleId, appearance)
+      let result = closure(deviceId, bundleId, targetAppBundleId)
 
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
+      switch result {
+      case let .success(output):
+        return .ok(.text(output))
 
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
+      case let .failure(error):
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
+  }
 
-    func onTriggerICloudSync(_ closure: @escaping (UUID, String?) -> Result<String, Swift.Error>) {
-        server.GET[ServerPath.iCloudSync.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
+  /// Callback to be executed on open url request.
+  /// - Parameter closure: The closure to be executed.
+  func onOpenUrl(_ closure: @escaping (UUID, String?, URL) -> Result<String, Swift.Error>) {
+    server.POST[ServerPath.openURL.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
 
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
+      guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
+        return .badRequest(.text("Bundle Id missing or corrupt."))
+      }
 
-            let result = closure(deviceId, bundleId)
+      let bodyData = Data(request.body)
+      let urlContainer: URLContainer
 
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
+      do {
+        urlContainer = try JSONDecoder().decode(URLContainer.self, from: bodyData)
+      } catch {
+        return .badRequest(.text(error.localizedDescription))
+      }
 
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
+      let result = closure(deviceId, bundleId, urlContainer.url)
+
+      switch result {
+      case let .success(output):
+        return .ok(.text(output))
+
+      case let .failure(error):
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
+  }
 
-    /// Callback to be executed on uninstall app device request.
-    /// - Parameter closure: The closure to be executed.
-    func onUninstallApp(_ closure: @escaping (UUID, String?, String) -> Result<String, Swift.Error>) {
-        server.GET[ServerPath.uninstallApp.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
+  func onGetAppContainer(_ closure: @escaping (UUID, String, AppContainer) -> Result<String, Swift.Error>) {
+    server.POST[ServerPath.getAppContainer.rawValue] = { request in
+      guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
+        return .badRequest(.text("Device Udid missing or corrupt."))
+      }
 
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
+      guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
+        return .badRequest(.text("Bundle Id missing or corrupt."))
+      }
 
-            guard let targetAppBundleId: String = request.headerValue(for: .targetBundleIdentifier) else {
-                return .badRequest(.text("No target app bundle id parameter provided."))
-            }
+      let bodyData = Data(request.body)
 
-            let result = closure(deviceId, bundleId, targetAppBundleId)
+      let appContainer: AppContainer
+      do {
+        appContainer = try JSONDecoder().decode(AppContainer.self, from: bodyData)
+      } catch {
+        return .badRequest(.text(error.localizedDescription))
+      }
 
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
+      let result = closure(deviceId, bundleId, appContainer)
 
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
+      switch result {
+      case let .success(output):
+        return .ok(.text(output))
+
+      case let .failure(error):
+        return .badRequest(.text(error.localizedDescription))
+      }
     }
-
-    /// Callback to be executed on set status bar override request.
-    /// - Parameter closure: The closure to be executed.
-    func onSetStatusBarOverride(_ closure: @escaping (UUID, String?, Set<StatusBarOverride>) -> Result<String, Swift.Error>) {
-        server.POST[ServerPath.statusBarOverrides.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
-
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
-
-            let bodyData = Data(request.body)
-            let decoder = JSONDecoder()
-            do {
-                let overrides: Set<StatusBarOverride> = try decoder.decode(Set<StatusBarOverride>.self, from: bodyData)
-
-                let result = closure(deviceId, bundleId, overrides)
-
-                switch result {
-                case let .success(output):
-                    return .ok(.text(output))
-
-                case let .failure(error):
-                    return .badRequest(.text(error.localizedDescription))
-                }
-            } catch {
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
-    }
-
-    /// Callback to be executed on clear status bar override request.
-    /// - Parameter closure: The closure to be executed.
-    func onClearStatusBarOverrides(_ closure: @escaping (UUID, String?) -> Result<String, Swift.Error>) {
-        server.GET[ServerPath.statusBarOverrides.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
-
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
-
-            let result = closure(deviceId, bundleId)
-
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
-
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
-    }
-
-    /// Callback to be executed on open url request.
-    /// - Parameter closure: The closure to be executed.
-    func onOpenUrl(_ closure: @escaping (UUID, String?, URL) -> Result<String, Swift.Error>) {
-        server.POST[ServerPath.openURL.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
-
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
-
-            let bodyData = Data(request.body)
-            let urlContainer: URLContainer
-
-            do {
-                urlContainer = try JSONDecoder().decode(URLContainer.self, from: bodyData)
-            } catch {
-                return .badRequest(.text(error.localizedDescription))
-            }
-
-            let result = closure(deviceId, bundleId, urlContainer.url)
-
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
-
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
-    }
-
-    func onGetAppContainer(_ closure: @escaping (UUID, String, AppContainer) -> Result<String, Swift.Error>) {
-        server.POST[ServerPath.getAppContainer.rawValue] = { request in
-            guard let deviceId = request.headerValue(for: .deviceUdid, UUID.init) else {
-                return .badRequest(.text("Device Udid missing or corrupt."))
-            }
-
-            guard let bundleId = request.headerValue(for: .bundleIdentifier) else {
-                return .badRequest(.text("Bundle Id missing or corrupt."))
-            }
-
-            let bodyData = Data(request.body)
-
-            let appContainer: AppContainer
-            do {
-                appContainer = try JSONDecoder().decode(AppContainer.self, from: bodyData)
-            } catch {
-                return .badRequest(.text(error.localizedDescription))
-            }
-
-            let result = closure(deviceId, bundleId, appContainer)
-
-            switch result {
-            case let .success(output):
-                return .ok(.text(output))
-
-            case let .failure(error):
-                return .badRequest(.text(error.localizedDescription))
-            }
-        }
-    }
+  }
 }
