+++ Sources/SimctlShared/DeviceAppearance.swift
@@ -1,0 +1,7 @@
+
+public enum DeviceAppearance: String {
+  /// The Light appearance style.
+  case light
+  /// The Dark appearance style.
+  case dark
+}
