+++ Sources/SimctlShared/URLContainer.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public struct URLContainer: Codable {
+  public let url: URL
+
+  public init(url: URL) {
+    self.url = url
+  }
+}
